+++ lib/presentation/component/recipe_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/presentation/component/app_image.dart';
-import 'package:recipe_app/presentation/component/skelton_animtion_widget.dart';
+import 'package:recipe_app/presentation/component/image_component/app_image.dart';
+import 'package:recipe_app/presentation/component/image_component/skelton_animtion_widget.dart';
 import 'package:recipe_app/ui/color_style.dart';
 import 'package:recipe_app/ui/text_style.dart';
